+++ R/Lib_Validation_biodivMapR.R
@@ -24,8 +24,7 @@
 #' @param FDmetric character. Functional diversity metric
 #'
 #' @return alpha and beta diversity metrics
-#' @importFrom raster raster projection compareCRS
-#' @importFrom rgdal readOGR
+#' @importFrom terra rast vect same.crs
 #' @importFrom tools file_path_sans_ext
 #' @export
 diversity_from_plots = function(Raster_SpectralSpecies, Plots, nbclusters = 50,
@@ -43,42 +42,35 @@ diversity_from_plots = function(Raster_SpectralSpecies, Plots, nbclusters = 50,
   # initialize alpha diversity for each plot
   Richness <- Shannon <- Fisher <- Simpson <- data.frame()
   Name.Vector <- list()
-  # prepare directory to write shapefiles with correct projection
-  Dir.Vector <- dirname(Plots[[1]])
-  Dir.Vector.reproject <- paste(Dir.Vector,'/Reproject',sep='')
 
   ## ______________________________________________________##
   ###       Read pixel coordinates for each polygon       ###
   ## ______________________________________________________##
-# total number of plots among all files
-  nbPolygons <-0
+
   # list of corodinates corresponding to each polygon to be studied
+  Raster <- rast(Raster_SpectralSpecies, lyrs=1)
   XY <- list()
   for (ip in 1:nbPlots){
-    # prepare for possible reprojection
     File.Vector <- Plots[[ip]]
     Name.Vector[[ip]] <- tools::file_path_sans_ext(basename(File.Vector))
     print(paste('Reading pixels coordinates for polygons in ',Name.Vector[[ip]],sep=''))
-    File.Vector.reproject <- paste(Dir.Vector.reproject,'/',Name.Vector[[ip]],'.shp','sep'='')
 
     if (file.exists(paste(tools::file_path_sans_ext(File.Vector),'.shp',sep=''))){
-      Plot <- rgdal::readOGR(Dir.Vector,Name.Vector[[ip]],verbose = FALSE)
+      Plot <- terra::vect(Plots[[ip]])
       # check if vector and rasters are in the same referential
       # if not, convert vector file
-      if (!raster::compareCRS(raster(Raster_SpectralSpecies), Plot)){
+      if (!same.crs(Raster, Plot)){
         stop('Raster and Plots have different projection. Plots should be reprojected to Raster CRS')
       }
     } else if (file.exists(paste(File.Vector,'kml','sep'='.'))){
       print('Please convert vector file to shpfile')
     }
     # extract data corresponding to the Raster_SpectralSpecies
-    XY0 <- extract_pixels_coordinates.From.OGR(Raster_SpectralSpecies,Plot)
-    # add each polygon in the shapefile to the XY list
-    for (ii in 1:length(XY0)){
-      nbPolygons <- nbPolygons+1
-      XY[[nbPolygons]] <- XY0[[ii]]
-    }
+    XY0 <- extract_pixels_coordinates(Raster,Plot)
+    XY <- c(XY, XY0)
   }
+  # total number of plots among all files
+  nbPolygons <- length(XY)
   message(paste('Number of polygons: ',nbPolygons))
 
   ## ______________________________________________________##
@@ -88,7 +80,7 @@ diversity_from_plots = function(Raster_SpectralSpecies, Plots, nbclusters = 50,
   Pixel_Inventory_All <- Pixel_Hellinger_All <- list()
   for (ip in 1:nbPolygons){
     # if only one polygon in the shapefile and if the polyon is not included in the Raster_SpectralSpecies
-    if (length(XY[[ip]]$col)==0){
+    if (all(is.na(XY[[ip]]$col))){
       # if list of individual plots provided
       if (length(Name_Plot)==nbPolygons){
         message(paste('Polygon named',Name_Plot[ip],'is out of the raster'))
@@ -172,7 +164,7 @@ diversity_from_plots = function(Raster_SpectralSpecies, Plots, nbclusters = 50,
     # for each polygon
     for (ip in 1:nbPolygons){
       # if only one polygon in the shapefile and if the polyon is not included in the raster
-      if (length(XY[[ip]]$col)==0){
+      if (all(is.na(XY[[ip]]$col))){
         FunctionalDiversity$FRic[ip] <- NA
         FunctionalDiversity$FEve[ip] <- NA
         FunctionalDiversity$FDiv[ip] <- NA
@@ -273,6 +265,10 @@ diversity_from_plots = function(Raster_SpectralSpecies, Plots, nbclusters = 50,
               'Name_Plot' = Name_Plot))
 }
 
+
+
+
+
 #' Get list of shapefiles in a directory
 #'
 #' @param x character or list. Directory containing shapefiles
@@ -292,57 +288,29 @@ list_shp <- function(x){
   return(List.Shp)
 }
 
-#' Extracts pixels coordinates from raster corresponding to an area defined by a vector
-#' @param Path.Raster path for the raster file. !! BIL expected
-#' @param Path.Vector path for the vector file. !! SHP expected
-#'
-#' @return ColRow list of coordinates of pixels corresponding to each polygon in shp
-#' @importFrom rgdal readOGR
-#' @importFrom tools file_path_sans_ext
-#' @importFrom raster raster cellFromPolygon
+
+#' Extracts pixels coordinates from raster intersecting a vector.
+#' @param x terra::SpatRaster or character. Raster or raster path.
+#' @param y terra::SpatVector or character. Vector or vector path.
+#' @return data.frame. Pixel indices row, col of intersecting parts.
+#' @import magrittr
+#' @importFrom terra rast vect cells rowFromCell colFromCell
+#' @importFrom dplyr mutate select group_by group_split
 #' @export
-
-extract_pixels_coordinates = function(Path.Raster,Path.Vector){
-  # read vector file
-  Shp_Path <- dirname(Path.Vector)
-  Shp_Name <- tools::file_path_sans_ext(basename(Path.Vector))
-  Shp_Crop <- rgdal::readOGR(Shp_Path,Shp_Name)
+extract_pixels_coordinates <- function(x, y){
   # read raster info
-  Raster <- raster::raster(Path.Raster, band = 1)
-  # extract pixel coordinates from raster based on vector
-  XY <- raster::cellFromPolygon (Raster,Shp_Crop)
-  # for each polygon in the
-  ColRow <- list()
-  for (i in 1:length(XY)){
-    ColRow[[i]] <- ind2sub(Raster,XY[[i]])
-  }
-  return(ColRow)
-}
+  if(is.character(x))
+    x <- terra::rast(x, lyrs = 1)
 
-#' Extracts pixels coordinates from raster corresponding to an area defined by a vector
-#' @param Path.Raster path for the raster file. !! BIL expected
-#' @param OGR.Vector  OGR for the vector file obtained from readOGR
-#'
-#' @return ColRow list of coordinates of pixels corresponding to each polygon in shp
-#' @importFrom raster raster cellFromXY
-#' @export
+  if(is.character(y))
+    y <- terra::vect(y)
 
-extract_pixels_coordinates.From.OGR = function(Path.Raster,OGR.Vector){
-  # read raster info
-  Raster <- raster::raster(Path.Raster, band = 1)
-  # for each polygon or point in the shapefile
-  ColRow <- list()
-  # extract pixel coordinates from raster based on vector
-  if (OGR.Vector@class[1]=='SpatialPointsDataFrame'){
-    XY <- raster::cellFromXY (Raster,OGR.Vector)
-    ColRow[[1]] <- ind2sub(Raster,XY)
+  ColRow <- cells(x, y) %>% as.data.frame() %>%
+    dplyr::mutate(row = rowFromCell(x, cell),
+                  col = colFromCell(x, cell)) %>%
+    dplyr::select(ID, row, col) %>%
+    dplyr::group_by(ID) %>% dplyr::group_split(.keep=F)
 
-  } else if  (OGR.Vector@class[1]=='SpatialPolygonsDataFrame'){
-    XY <- raster::cellFromPolygon (Raster,OGR.Vector)
-    for (i in 1:length(XY)){
-      ColRow[[i]] <- ind2sub(Raster,XY[[i]])
-    }
-  }
   return(ColRow)
 }
 
@@ -361,53 +329,3 @@ get_alpha_metrics = function(Distrib){
   return(list("Richness" = RichnessPlot,"fisher"=FisherPlot,"Shannon"=ShannonPlot,"Simpson"=SimpsonPlot))
 }
 
-#' #' build a vector file from raster footprint
-#' #' borrowed from https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/
-#' #' @param x path for a raster or raster object
-#' #' @param outshape path for a vector to be written
-#' #' @param gdalformat character. gdal format for shapefile
-#' #' @param pypath character. path for python
-#' #' @param readpoly boolean. should polygons be read once produced?
-#' #' @param quiet boolean. set T to avoid verbose
-#' #'
-#' #' @return NULL
-#' #' @importFrom rgdal readOGR
-#' #' @importFrom raster writeRaster
-#' #' @importFrom methods is
-#' gdal_polygonizeR = function(x, outshape=NULL, gdalformat = 'ESRI Shapefile',
-#'                             pypath=NULL, readpoly=TRUE, quiet=TRUE) {
-#'
-#'   if (is.null(pypath)) {
-#'     pypath <- Sys.which('gdal_polygonize.py')
-#'   }
-#'   if (!file.exists(pypath)) stop("Can't find gdal_polygonize.py on your system.")
-#'   owd <- getwd()
-#'   on.exit(setwd(owd))
-#'   setwd(dirname(pypath))
-#'   if (!is.null(outshape)) {
-#'     outshape  = sub('\\.shp$', '', outshape)
-#'     f.exists  = file.exists(paste(outshape, c('shp', 'shx', 'dbf'), sep='.'))
-#'     if (any(f.exists)){
-#'       print('Footprint of raster file already defined')
-#'     }
-#'     # stop(sprintf('File already exists: %s',
-#'     #              toString(paste(outshape, c('shp', 'shx', 'dbf'),
-#'     #                             sep='.')[f.exists])), call.=FALSE)
-#'   } else outshape <- tempfile()
-#'   if (is(x, 'Raster')) {
-#'     raster::writeRaster(x, {f <- tempfile(fileext='.tif')})
-#'     rastpath <- normalizePath(f)
-#'   } else if (is.character(x)) {
-#'     rastpath <- normalizePath(x)
-#'   } else stop('x must be a file path (character string), or a Raster object.')
-#'   outshape.Shp = paste(outshape,'.shp',sep='')
-#'   if (!file.exists(outshape.Shp)){
-#'     system2('python', args=(sprintf('"%1$s" "%2$s" -f "%3$s" "%4$s.shp"',
-#'                                     pypath, rastpath, gdalformat, outshape)))
-#'   }
-#'   if (readpoly==TRUE) {
-#'     shp <- rgdal::readOGR(dirname(outshape), layer = basename(outshape), verbose=!quiet)
-#'     return(shp)
-#'   }
-#'   return(NULL)
-#' }