+++ NEWS.md
@@ -1,3 +1,17 @@
+# biodivMapR v1.12.0
+
+# Fix
+
+- issue #18: remove the dependency to rgdal and rgeos using terra and gdalUtilities instead
+
+# Change
+
+- remove extract_pixels_coordinates.From.OGR, a unique extract_pixels_coordinates now works for both character or rast/vect inputs
+
+# Addition
+
+- add get_gdal_info to get raster gdal info as a nested list.
+
 # biodivMapR v1.11.1
 
 ## Change